requirements.txt
Django<1.8
MySQL-python==1.2.5
djangorestframework
markdown
django-filter